Refactoring, readme.
