+++ .devcontainer/devcontainer.json
@@ -2,8 +2,8 @@
     "image": "multiversx/development-playground:latest",
     "remoteEnv": {
         "CONTRACTS": "${containerWorkspaceFolder}/contracts",
-        "SNIPPETS_PY": "${containerWorkspaceFolder}/snippets/python",
-        "SNIPPETS_JS": "${containerWorkspaceFolder}/snippets/javascript",
+        "SNIPPETS_PY": "${containerWorkspaceFolder}/snippets-python",
+        "SNIPPETS_JS": "${containerWorkspaceFolder}/snippets-javascript",
         "SANDBOX": "${containerWorkspaceFolder}/sandbox",
         "TESTWALLETS": "/home/developer/multiversx-sdk/testwallets/latest/users"
     },
