+++ README.md
@@ -6,12 +6,12 @@ Playground for contracts and dApps.
 
 GitHub Codespaces are fully configured development environments in the cloud.
 
-[![Open in GitHub Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/multiversx/mx-sdk-playground/tree/init?quickstart=1)
+[![Open in GitHub Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/multiversx/mx-sdk-playground?quickstart=1)
 
 ## Open in Visual Studio Code Devcontainers
 
 If you already have **VS Code** and **Docker** installed, you can click on the badge below to get started. It will cause VS Code to automatically install the Dev Containers extension if needed, clone the source code into a container volume, and spin up a dev container for use.
 
-<a href="https://vscode.dev/redirect?url=vscode://ms-vscode-remote.remote-containers/cloneInVolume?url=https://github.com/multiversx/mx-sdk-playground/tree/init">
+<a href="https://vscode.dev/redirect?url=vscode://ms-vscode-remote.remote-containers/cloneInVolume?url=https://github.com/multiversx/mx-sdk-playground">
     <img alt="Open in Dev Containers" src="https://img.shields.io/static/v1?label=Dev%20Containers&message=Open&color=blue&logo=visualstudiocode" height="26px" />
 </a>
